+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### 4.0.2 (Next)
+* [#97](https://github.com/alexa-js/alexa-app/pull/97): Added support for custom directives - [@ajcrites](https://github.com/ajcrites).
 * Your contribution here
 
 ### 4.0.1 (May 17, 2017)